requirements.txt
streamlit
pypdf
python-docx
groq
openai
python-dotenv
